Simulate port is already allocated(chaos-mesh#62)
